refactor:  import path and global state
- update import paths
- consolidate global state management
