+++ apps/vscode-extension/src/utils/IsNewDayHandler.ts
@@ -1,14 +1,13 @@
-import { SYNC_DATA_KEY, filesData } from "../constants";
 import { GlobalStateData } from "../types-schemas";
-import { getExtensionContext } from "../extension";
+import { filesData } from "../constants";
 import getTodaysLocalDate from "@repo/common/getTodaysLocalDate";
+import updateGlobalStateData from "./global-state/updateGlobalStateData";
 
 const isNewDayHandler = async (
   dailyData: GlobalStateData["dailyData"],
   lastServerSync: Date,
 ) => {
   const todaysDateString = getTodaysLocalDate();
-  const context = getExtensionContext();
 
   if (!Object.hasOwn(dailyData, todaysDateString)) {
     Object.keys(filesData).forEach((key) => {
@@ -28,7 +27,7 @@ const isNewDayHandler = async (
       },
     };
 
-    await context.globalState.update(SYNC_DATA_KEY, newGlobalStateData);
+    await updateGlobalStateData(newGlobalStateData);
 
     return newGlobalStateData;
   }
